requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
chromadb
neo4j
py2neo
langchain-text-splitters
PyPDF2
python-docx
tqdm
networkx
openai
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
# If you use GraphRAG, add its package name here (if available)
